Map 1.3, Music Brick, Hidden Music Brick, Fly Goomba

+++ 05-ScenceManager/Animations.cpp
@@ -22,7 +22,7 @@ void CAnimation::Add(int spriteId, DWORD time)
 // NOTE: sometimes Animation object is NULL ??? HOW ??? 
 void CAnimation::Render(float x, float y, int alpha)
 {
-	DWORD now = GetTickCount64();
+	ULONGLONG now = GetTickCount64();
 	if (currentFrame == -1)
 	{
 		currentFrame = 0;
@@ -30,7 +30,7 @@ void CAnimation::Render(float x, float y, int alpha)
 	}
 	else
 	{
-		DWORD t = frames[currentFrame]->GetTime();
+		ULONGLONG t = frames[currentFrame]->GetTime();
 		if (now - lastFrameTime > t)
 		{
 			currentFrame++;
